Replace hard coded cof value with ones assigned during tile creation

+++ dynamic_object.h
@@ -77,8 +77,8 @@ typedef struct {
   float normalForce;
   float thrustX;
   float thrustY;
-  int frictionalForceX;
-  int frictionalForceY;
+  float frictionalForceX;
+  float frictionalForceY;
   int isMoving;
   int isRunning;
   int moveLeft;
